+++ src/main/kotlin/io/acari/n7/notification/UpdateComponent.kt
@@ -5,13 +5,13 @@ import com.intellij.openapi.project.Project
 
 import io.acari.n7.config.ConfigurationPersistence
 
-const val VERSION = "v1.6.0"
+const val VERSION = "v1.7.0"
 val UPDATE_MESSAGE =
     """
       What's New?<br>
       <ul>
-      <li>Backgrounds can be made transparent!</li>
-      <li>Report exceptions anonymously!</li>
+      <li>Enhanced Settings Menu.</li>
+      <li>Restored some theme integrations.</li>
       <br>
       Thanks again for downloading <b>Normandy Progress Bar UI</b>! •‿•<br>
           <br>See <a href="https://github.com/cyclic-reference/normandy-progress-bar/blob/master/docs/CHANGELOG.md">Changelog</a> for more details.