+++ src/main/kotlin/io/acari/n7/ui/GuidanceSystem.kt
@@ -1,6 +1,6 @@
 package io.acari.n7.ui
 
-import com.intellij.util.ui.JBUI.*
+import com.intellij.ui.scale.JBUIScale.scale
 
 const val DEFAULT_VELOCITY_FROM_CITADEL = 2f
 const val DEFAULT_DISTANCE_FROM_CITADEL = 0f
@@ -37,12 +37,21 @@ class GuidanceSystem {
         val distanceBetweenCitadelAndNormandy = distanceFromCitadel
         val headingToCitadel = isHeadingToCitadel()
         val whiteSpaceOffset = scale(2f)//the reversed normandy has a bit of whitespace in the front so when it is flipped the nozzles do not match the contrails
-        val startingX = if (headingToCitadel) { distanceBetweenCitadelAndNormandy + iconWidth - whiteSpaceOffset
-        } else 2f * offset
-        val distanceBetweenNormandyAndOmega = componentWidth - distanceBetweenCitadelAndNormandy
-        val lengthOfContrail = if (headingToCitadel) distanceBetweenNormandyAndOmega else distanceBetweenCitadelAndNormandy - iconWidth
-        val positionOfNormandy = if (headingToCitadel) distanceBetweenCitadelAndNormandy + whiteSpaceOffset else distanceBetweenCitadelAndNormandy - iconWidth
-        NormandyPositionData(startingX, lengthOfContrail, positionOfNormandy.toInt())
+        val startingX =
+            if (headingToCitadel) distanceBetweenCitadelAndNormandy + iconWidth - whiteSpaceOffset
+            else 2f * offset
+        val lengthOfContrail =
+            if (headingToCitadel) componentWidth - distanceBetweenCitadelAndNormandy - iconWidth
+            else distanceBetweenCitadelAndNormandy - iconWidth
+        val positionOfNormandy =
+            if (headingToCitadel) distanceBetweenCitadelAndNormandy
+            else distanceBetweenCitadelAndNormandy - iconWidth
+
+        NormandyPositionData(
+            startingX,
+            lengthOfContrail,
+            positionOfNormandy.toInt()
+        )
       }
 
   private fun isAtOmega(componentWidth: Int) = distanceFromCitadel >= componentWidth + iconWidth